test(db): move schema snapshot checks from base_cqrs
